Fix 43 compiler warnings across C++ and C# codebase
C++ Core:
- Add [[maybe_unused]] for literalMode (LynxPpuTools) and bppMask (LynxSuzy)
- Replace strncpy with strncpy_s (LynxConsole)

C++ Tests:
- Remove unused variables (initialAcc, testWord, value)
- Use unused bpp2/bpp3/timer2Static variables in tests
- Add [[maybe_unused]] for unreferenced prgSize parameter
- Fix int8_t truncation cast with signed constant
- Add C4310 pragma for intentional hardware register casts
- Capture [[nodiscard]] return value in Trim test
- Cast (void) to suppress nodiscard warning

C++ Benchmarks:
- Add [[maybe_unused]] for GenerateNoise and iFlag
- Fix int8_t Volume overflow (use valid max value 127)

C++ Utilities:
- Add static_cast for narrowing conversions (Base64, FastString,
  PNGHelper, Serializer, Socket, AviWriter)
- Fix signed/unsigned mismatch (~0 to ~0u in AviWriter)
- Add [[maybe_unused]] for unreferenced parameters
- Fix Serializer double-to-T cast

C# UI:
- Replace obsolete GetGlyphSpan() with Glyphs (5 locations)
- Suppress CS0067 for reserved MarkerClicked event
- Suppress CS0618 for FilterQuality (TODO: migrate to SKSamplingOptions)
- Suppress CS8500 for required P/Invoke interop pointer
- Fix CS8604 null reference with null-coalescing fallback

All 1495 tests pass. Zero warnings from our code remain.

+++ Utilities/Base64.h
@@ -64,7 +64,7 @@ public:
 			val = (val << 6) + T[c];
 			valb += 6;
 			if (valb >= 0) {
-				out.push_back(val >> valb);
+				out.push_back(static_cast<uint8_t>(val >> valb));
 				valb -= 8;
 			}
 		}
